+++ src/Radical.Windows.Tests/Conventions/GenerateViewModelStaticResourceKeyTests.cs
@@ -2,7 +2,7 @@
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using Radical.ComponentModel.Messaging;
 using Radical.Windows.ComponentModel;
-using Radical.Windows.Presentation.Boot;
+using Radical.Windows.Bootstrap;
 using Radical.Windows.Services;
 
 namespace Radical.Windows.Tests.Conventions
