+++ src/Radical.Windows/Bootstrap/ApplicationBootstrapper(generic).cs
@@ -1,4 +1,4 @@
-﻿namespace Radical.Windows.Presentation.Boot
+﻿namespace Radical.Windows.Bootstrap
 {
     public class ApplicationBootstrapper<TMainView> : ApplicationBootstrapper where TMainView : System.Windows.Window
     {
