+++ src/Radical.Windows/Bootstrap/DefaultComponentReleaser.cs
@@ -1,6 +1,6 @@
 ﻿using Radical.Windows.ComponentModel;
 
-namespace Radical.Windows.Presentation.Boot
+namespace Radical.Windows.Bootstrap
 {
     class DefaultComponentReleaser : IReleaseComponents
     {
