+++ src/Radical.Windows/Bootstrap/Features/AutoRegionsInjection.cs
@@ -4,7 +4,7 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 
-namespace Radical.Windows.Presentation.Boot.Features
+namespace Radical.Windows.Bootstrap.Features
 {
     class AutoRegionsInjection : IFeature
     {
