+++ src/Radical.Windows/Bootstrap/Installers/MessagingInstaller.cs
@@ -2,12 +2,12 @@
 using Radical.ComponentModel.Messaging;
 using Radical.Linq;
 using Radical.Messaging;
-using Radical.Windows.Presentation.Boot.Features;
+using Radical.Windows.Bootstrap.Features;
 using System;
 using System.Collections.Generic;
 using System.Linq;
 
-namespace Radical.Windows.Presentation.Boot.Installers
+namespace Radical.Windows.Bootstrap.Installers
 {
     class MessagingInstaller : IDependenciesInstaller
     {
