+++ src/Radical.Windows/Services/ConventionsHandler.cs
@@ -5,7 +5,7 @@ using Radical.Linq;
 using Radical.Reflection;
 using Radical.Windows.Behaviors;
 using Radical.Windows.ComponentModel;
-using Radical.Windows.Presentation.Boot;
+using Radical.Windows.Bootstrap;
 using Radical.Windows.Regions;
 using System;
 using System.Linq;
